+++ tests/testthat.R
@@ -1,9 +1,4 @@
 library(testthat)
 library(easycensus)
 
-check_api = function() {
-    if (Sys.getenv("CENSUS_API_KEY") == "")
-        skip("Census API not available")
-}
-
 test_check("easycensus")