+++ tests/testthat/test-find_table.R
@@ -1,4 +1,4 @@
 test_that("tables are found for common use cases", {
-    expect_equal(find_dec_table("sex", "age", show=-1), "P012")
-    expect_equal(find_dec_table("hisp", show=-1), "P004")
+    expect_equal(cens_find_dec("sex", "age", show=-1), "P12")
+    expect_equal(cens_find_dec("hisp", show=-1), "P4")
 })