+++ tests/testthat/test-tidiers.R
@@ -1,10 +1,10 @@
 test_that("tidiers produce no warnings", {
   expect_warning({
-      tidy_age(tables_sf1$P012$vars$age)
-      tidy_age_bins(tables_sf1$P012$vars$age)
-      tidy_age_bins(tables_sf1$P012$vars$age, as_factor=TRUE)
-      tidy_race(tables_sf1$P012$vars$race_ethnicity)
-      tidy_race(tables_sf1$P008$vars$race_1)
+      tidy_age(tables_sf1$P12$vars$age)
+      tidy_age_bins(tables_sf1$P12$vars$age)
+      tidy_age_bins(tables_sf1$P12$vars$age, as_factor=TRUE)
+      tidy_race(tables_sf1$P12$vars$race_ethnicity)
+      tidy_race(tables_sf1$P8$vars$race_1)
   }, regex=NA) # aka expect _no_ warning
 })